refactor src.djai.util.cli.run_cmd with context managers

+++ src/djai/util/cli/run_cmd.py
@@ -25,67 +25,75 @@ _WSGI_PY_FILE_SRC_PATH = _SERVER_FILES_DIR_PATH / _WSGI_PY_FILE_NAME
 _PROCFILE_NAME = 'Procfile'
 
 
-class HandleManagePy(AbstractContextManager):
+class ManagePyHandling(AbstractContextManager):
     """Handle `manage.py` file."""
 
     def __enter__(self):
         """Add `manage.py` file, if applicable."""
         # pylint: disable=attribute-defined-outside-init
-        self._manage_py_missing = not Path(_MANAGE_PY_FILE_NAME).exists()
+        self.manage_py_missing: bool = not Path(_MANAGE_PY_FILE_NAME).exists()
 
-        if self._manage_py_missing:
+        if self.manage_py_missing:
             shutil.copyfile(src=_MANAGE_PY_FILE_SRC_PATH,
                             dst=_MANAGE_PY_FILE_NAME,
                             follow_symlinks=False)
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         """Remove `manage.py` file, if applicable."""
-        if self._manage_py_missing:
+        if self.manage_py_missing:
             os.remove(path=_MANAGE_PY_FILE_NAME)
             assert not Path(_MANAGE_PY_FILE_NAME).exists()
 
 
-class HandleServerFiles(AbstractContextManager):
+class SGIFilesHandling(AbstractContextManager):
     """Handle `asgi.py`/`wsgi.py` file & Procfile."""
 
     def __init__(self, asgi: Optional[str] = None):
         """Initialize context manager."""
-        self.asgi = asgi
+        self.asgi: Optional[str] = asgi
 
     def __enter__(self):
         """Add `asgi.py`/`wsgi.py` file & Procfile, if applicable."""
+        # pylint: disable=attribute-defined-outside-init
         if self.asgi:
-            assert not Path(_ASGI_PY_FILE_NAME).exists()
-            shutil.copyfile(src=_ASGI_PY_FILE_SRC_PATH,
-                            dst=_ASGI_PY_FILE_NAME,
-                            follow_symlinks=False)
+            self.agsi_py_missing: bool = not Path(_ASGI_PY_FILE_NAME).exists()
+            if self.agsi_py_missing:
+                shutil.copyfile(src=_ASGI_PY_FILE_SRC_PATH,
+                                dst=_ASGI_PY_FILE_NAME,
+                                follow_symlinks=False)
 
-            assert not Path(_PROCFILE_NAME).exists()
-            shutil.copyfile(src=(_SERVER_FILES_DIR_PATH /
-                                 f'{_PROCFILE_NAME}.{self.asgi.capitalize()}'),
-                            dst=_PROCFILE_NAME,
-                            follow_symlinks=False)
+            self.procfile_missing: bool = not Path(_PROCFILE_NAME).exists()
+            if self.procfile_missing:
+                shutil.copyfile(
+                    src=(_SERVER_FILES_DIR_PATH /
+                         f'{_PROCFILE_NAME}.{self.asgi.capitalize()}'),
+                    dst=_PROCFILE_NAME,
+                    follow_symlinks=False)
 
         else:
-            assert not Path(_WSGI_PY_FILE_NAME).exists()
-            shutil.copyfile(src=_WSGI_PY_FILE_SRC_PATH,
-                            dst=_WSGI_PY_FILE_NAME,
-                            follow_symlinks=False)
+            self.wsgi_py_missing: bool = not Path(_WSGI_PY_FILE_NAME).exists()
+            if self.wsgi_py_missing:
+                shutil.copyfile(src=_WSGI_PY_FILE_SRC_PATH,
+                                dst=_WSGI_PY_FILE_NAME,
+                                follow_symlinks=False)
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         """Remove `asgi.py`/`wsgi.py` file & Procfile, if applicable."""
         if self.asgi:
-            os.remove(path=_ASGI_PY_FILE_NAME)
-            assert not Path(_ASGI_PY_FILE_NAME).exists()
-            os.remove(path=_PROCFILE_NAME)
-            assert not Path(_PROCFILE_NAME).exists()
+            if self.asgi_py_missing:
+                os.remove(path=_ASGI_PY_FILE_NAME)
+                assert not Path(_ASGI_PY_FILE_NAME).exists()
 
-        else:
+            if self.procfile_missing:
+                os.remove(path=_PROCFILE_NAME)
+                assert not Path(_PROCFILE_NAME).exists()
+
+        elif self.wsgi_py_missing:
             os.remove(path=_WSGI_PY_FILE_NAME)
             assert not Path(_WSGI_PY_FILE_NAME).exists()
 
 
-class HandleGitHash(AbstractContextManager):
+class GitHashHandling(AbstractContextManager):
     """Handle `Procfile` file."""
 
     def __enter__(self):
@@ -93,7 +101,7 @@ class HandleGitHash(AbstractContextManager):
         assert not Path(_GIT_HASH_FILE_NAME).exists()
 
         # pylint: disable=attribute-defined-outside-init
-        self.git_hash = get_git_repo_head_commit_hash()
+        self.git_hash: Optional[str] = get_git_repo_head_commit_hash()
 
         if self.git_hash:
             with open(file=_GIT_HASH_FILE_NAME,
@@ -116,5 +124,5 @@ class HandleGitHash(AbstractContextManager):
 def run_cmd(command: str,
             asgi: Optional[str] = None):
     """Run a certain CLI command."""
-    with HandleManagePy(), HandleServerFiles(asgi=asgi), HandleGitHash():
+    with ManagePyHandling(), SGIFilesHandling(asgi=asgi), GitHashHandling():
         os.system(command=command)
